chore: refactor after file removal
