--- packages/path-kaltura/src/utils.ts
@@ -1,4 +1,0 @@
-
-
-
-
